* compilation bug corrected * but it seems to remain an execution bug
git-svn-id: file:///Users/ng/svn-local/to_share/frb/trunk@59 f04dad46-9b18-0410-952a-bb25a776dc73

+++ src/parser/frbfunction.h
@@ -288,7 +288,7 @@ inline FrBCodeFunction::VarID FrBCodeFunction::findLocalVar(String name) const
 
 inline const FrBClass * FrBCodeFunction::getLocalVar(FrBCodeFunction::VarID id) const
 {
-    frb_assert(id >= 0 && id < _var.size());
+    frb_assert(id >= 0 && id < (int)_var.size());
     return _var[id];
 }
